Renamed some classes to better fit new changes
